Centralise and validate item viewer types against IIIF specs
