+++ content/webapp/types/manifest.ts
@@ -2,7 +2,6 @@ import type {
   ChoiceBody,
   CollectionItems,
   ContentResource,
-  InternationalString,
   Manifest,
   MetadataItem,
   Range,
@@ -30,12 +29,6 @@ export type ServiceWithMetadata = Service & {
 import type { TransformedAuthService } from '@weco/content/utils/iiif/v3';
 
 export type ThumbnailImage = { url: string; width: number };
-
-export type Original = {
-  originalFile: string | undefined;
-  label: InternationalString | null | undefined;
-  format: string | undefined;
-};
 
 export type TransformedCanvas = {
   id: string;
@@ -100,3 +93,14 @@ export type CustomSpecificationBehaviors =
 export type CustomContentResource = ContentResource & {
   behavior: 'original';
 } & { type?: string; format?: string };
+
+// Some of our ContentResources can have a type of 'Audio':
+// https://iiif.wellcomecollection.org/presentation/v3/b17276342
+// However, the IIIF Presentation API spec only has a type of 'Sound'
+// so we add 'Audio' to the type here.
+export type IIIFItemProps =
+  | (Omit<ContentResource, 'type'> & {
+      type: ContentResource['type'] | 'Audio';
+    })
+  | CustomContentResource
+  | ChoiceBody;
